Move available_monitors and primary_monitor to EventLoopWindowTarget (#1616)
