+++ src/platform_impl/macos/event_loop.rs
@@ -34,6 +34,18 @@ impl<T> Default for EventLoopWindowTarget<T> {
     }
 }
 
+impl<T: 'static> EventLoopWindowTarget<T> {
+    #[inline]
+    pub fn available_monitors(&self) -> VecDeque<MonitorHandle> {
+        monitor::available_monitors()
+    }
+
+    #[inline]
+    pub fn primary_monitor(&self) -> MonitorHandle {
+        monitor::primary_monitor()
+    }
+}
+
 pub struct EventLoop<T: 'static> {
     window_target: Rc<RootWindowTarget<T>>,
     _delegate: IdRef,
@@ -66,16 +78,6 @@ impl<T> EventLoop<T> {
             }),
             _delegate: delegate,
         }
-    }
-
-    #[inline]
-    pub fn available_monitors(&self) -> VecDeque<MonitorHandle> {
-        monitor::available_monitors()
-    }
-
-    #[inline]
-    pub fn primary_monitor(&self) -> MonitorHandle {
-        monitor::primary_monitor()
     }
 
     pub fn window_target(&self) -> &RootWindowTarget<T> {
